Inject function `swup.getFragmentVisit(route)`
