+++ src/inc/functions.ts
@@ -1,6 +1,6 @@
 import { Location } from 'swup';
 import type { Visit } from 'swup';
-import { default as FragmentPlugin } from '../SwupFragmentPlugin.js';
+import SwupFragmentPlugin, { default as FragmentPlugin } from '../SwupFragmentPlugin.js';
 import type { Rule, Route, FragmentVisit, FragmentElement } from '../SwupFragmentPlugin.js';
 import Logger, { highlight } from './Logger.js';
 
@@ -330,4 +330,26 @@ export function shouldSkipAnimation({ swup }: FragmentPlugin): boolean {
  */
 export function dedupe<T>(arr: Array<T>): Array<T> {
 	return [...new Set<T>(arr)];
+}
+
+/**
+ * Get the fragment visit object for a given route
+ */
+export function getFragmentVisit(this: SwupFragmentPlugin, route: Route, logger?: Logger): FragmentVisit | undefined {
+	const rule = getFirstMatchingRule(route, this.rules);
+
+	// Bail early if no rule matched
+	if (!rule) return;
+
+	// Validate the containers from the matched rule
+	const containers = getFragmentsForVisit(route, rule.containers, logger);
+	// Bail early if there are no valid containers for the rule
+	if (!containers.length) return;
+
+	const visit: FragmentVisit = {
+		rule,
+		containers
+	};
+
+	return visit;
 }
